seperated main component into smaller ones
