+++ client/pages/format-time.js
@@ -1,0 +1,9 @@
+import { fromUnixTime, format } from 'date-fns';
+const formatTime = props => {
+
+  const unix = fromUnixTime(props);
+  const gameTime = format(unix, 'KK:mm bb');
+  return gameTime;
+
+};
+export default formatTime;
