+++ client/pages/home.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Header from './header';
-import FixtureList from './fixtureList';
+import FixtureList from './fixture-list';
 import Footer from './footer';
 export default function Home(props) {
   return (
