+++ client/pages/no-games-today.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const NoGamesToday = () => {
+  return (
+      <div className="row">
+        <div className="no-game">
+          <h1>No Upcoming Fixtures</h1>
+        </div>
+      </div>
+  );
+
+};
+export default NoGamesToday;
