+++ client/pages/today-fixtures.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const TodayFixtures = () => {
+  return (
+  <div className = "fixture-date-heading">
+       <h1>Today&apos;s Fixtures</h1>
+        </div>
+  );
+
+};
+
+export default TodayFixtures;
